move other db functions into database.ts
