+++ src/server/getters/get-markets-awaiting-reporting.ts
@@ -1,7 +1,6 @@
 import * as Knex from "knex";
 import { Address, MarketsContractAddressRow, UIMarketInfo, UIMarketsInfo, ErrorCallback } from "../../types";
-import { reshapeMarketsRowToUIMarketInfo, getMarketsWithReportingState } from "./get-market-info";
-import { queryModifier } from "./database";
+import { queryModifier, reshapeMarketsRowToUIMarketInfo, getMarketsWithReportingState } from "./database";
 
 // Look up all markets that are currently available for limited reporting.
 // Must be able to sort by number of reports submitted for each market so far, and the response should include the number of reports already submitted -- as well as the payoutNumerator values of each of the reports + the amount staked on each -- as part of the response.
